+++ src/java/ie/omk/smpp/util/SequenceNumberScheme.java
@@ -19,14 +19,14 @@ public interface SequenceNumberScheme {
      * Constant that can be returned from the <code>peek</code> methods to
      * indicate that the peek operation is not supported.
      */
-    int PEEK_UNSUPPORTED = -1;
+    long PEEK_UNSUPPORTED = -1;
 
     /**
      * Get the next number in this sequence's scheme. An implementation of this
      * interface <b>must </b> guard against multi-threaded access to this method
      * to prevent more than one thread getting the same sequence number.
      */
-    int nextNumber();
+    long nextNumber();
 
     /**
      * Get the next number in this sequence's scheme without causing it to move
@@ -35,16 +35,16 @@ public interface SequenceNumberScheme {
      * increasing the sequence. Multiple calls to <code>peek</code> will
      * return the same number until a call to <code>nextNumber</code> is made.
      */
-    int peek();
+    long peek();
 
     /**
      * Get the nth next number in this sequence's scheme without causing it to
      * move to the next-in-sequence. This method returns the <code>nth</code>
      * next number in the sequence. This is an optional operation. If a sequence
      * numbering scheme does not support this operation, it should always return
-     * {@link #PEEK_UNSUPPORTED}to the caller.
+     * {@link #PEEK_UNSUPPORTED} to the caller.
      */
-    int peek(int nth);
+    long peek(long nth);
 
     /**
      * Reset the sequence scheme to the beginning of the sequence.
